"added calculator"

+++ src/main/java/calculator/Remainder.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class Remainder {
 
-	public static double getResult(int num1, int num2) {
+	public static double getResult(double num1, double num2) {
 		return num1 % num2;
 		
 	}
